Improved README
Added the consolefont hook to the installation instruction for Arch
Linux, because with my new nvidia card I had trouble to clear the screen
entirely. Roughly one half of the screen kept the previous color. With
the correct order of these two (related) hooks, that problem is fixed.
Why does it always have to be so, that nvidia plays an extra role?